Changes binlog flushing and purging behavior (#24)
* Rotates binlog only if increment interval > 0

This avoids interference in the incremental backup
interval of several instances connected against the same DB

* Makes purging binlogs optional, Adds purge before

- binlogs are only purged on rotate if enabled
- binlogs can be purged with "PURGE BINLOGS BEFORE " to
keep them for a longer time (MariaDB 10.6 brings flag handling this)

* Makes purging enabled by default, adds readme

+++ pkg/database/mariadb.go
@@ -286,7 +286,9 @@ func (m *MariaDB) StartIncBackup(ctx context.Context, mp LogPosition, dir string
 	if err != nil {
 		return fmt.Errorf("Cannot start binlog stream: %w", err)
 	}
-	m.flushTimer = time.AfterFunc(time.Duration(m.cfg.Backup.IncrementalBackupInMinutes)*time.Minute, func() { m.flushLogs("") })
+	if m.cfg.Backup.IncrementalBackupInMinutes > 0 {
+		m.flushTimer = time.AfterFunc(time.Duration(m.cfg.Backup.IncrementalBackupInMinutes)*time.Minute, func() { m.flushLogs("") })
+	}
 	for {
 		ev, inerr := streamer.GetEvent(ctx)
 		if inerr != nil {
@@ -324,11 +326,11 @@ func (m *MariaDB) StartIncBackup(ctx context.Context, mp LogPosition, dir string
 
 		switch ev.Event.(type) {
 		case *replication.RowsEvent:
-			if m.flushTimer == nil {
+			if m.flushTimer == nil && m.cfg.Backup.IncrementalBackupInMinutes > 0 {
 				m.flushTimer = time.AfterFunc(time.Duration(m.cfg.Backup.IncrementalBackupInMinutes)*time.Minute, func() { m.flushLogs(binlogFile) })
 			}
 		case *replication.QueryEvent:
-			if m.flushTimer == nil {
+			if m.flushTimer == nil && m.cfg.Backup.IncrementalBackupInMinutes > 0 {
 				m.flushTimer = time.AfterFunc(time.Duration(m.cfg.Backup.IncrementalBackupInMinutes)*time.Minute, func() { m.flushLogs(binlogFile) })
 			}
 		}
@@ -382,8 +384,18 @@ func (m *MariaDB) flushLogs(binlogFile string) (err error) {
 		return
 	}
 
-	if binlogFile != "" {
-		return purgeBinlogsTo(m.cfg.Database, binlogFile)
+	if !m.cfg.Backup.DisableBinlogPurgeOnRotate {
+		if m.cfg.Backup.PurgeBinlogAfterMinutes > 0 {
+			err = purgeBinlogsBefore(m.cfg.Database, m.cfg.Backup.PurgeBinlogAfterMinutes)
+			if err != nil {
+				log.Warn("error purging binlogs: ", err)
+			}
+		} else if binlogFile != "" {
+			err = purgeBinlogsTo(m.cfg.Database, binlogFile)
+			if err != nil {
+				log.Warn("error purging binlogs: ", err)
+			}
+		}
 	}
 	return
 }
